requirements.txt
django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
drf-spectacular==0.26.5
